MindCare/MindCare AI v1/services/ai-pipeline/utils/auth.py
"""JWT token validation for API endpoints."""

from fastapi import HTTPException, Security
from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
from jose import JWTError, jwt

from models.config import get_settings

security = HTTPBearer()


async def verify_token(
    credentials: HTTPAuthorizationCredentials = Security(security),
) -> dict:
    """Validate JWT token from Supabase Auth."""
    settings = get_settings()
    token = credentials.credentials

    try:
        payload = jwt.decode(
            token,
            settings.jwt_secret,
            algorithms=[settings.jwt_algorithm],
            audience="authenticated",
        )
        user_id = payload.get("sub")
        if user_id is None:
            raise HTTPException(status_code=401, detail="Invalid token: no user ID")
        return {"user_id": user_id, "role": payload.get("role", "authenticated")}
    except JWTError as e:
        raise HTTPException(status_code=401, detail=f"Invalid token: {e}")
