MindCare/MindCare AI v1/services/ai-pipeline/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
deepgram-sdk==3.8.0
anthropic==0.40.0
openai==1.57.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
httpx==0.28.1
python-jose[cryptography]==3.3.0
structlog==24.4.0
tenacity==9.0.0
pytest==8.3.4
pytest-asyncio==0.24.0
